Model/requirements.txt
flask>=2.3.0
gunicorn>=21.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
statsmodels>=0.14.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
joblib>=1.3.0
